+++ prisma/migrations/20250704214021_changed_type_from_item_to_buy_in_order/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+UPDATE "Order" SET "type" = 'BUY' WHERE "type" = 'ITEM';